refactor: 💡 (KeywordsInMemoryRepository) getOrCreateValue 메서드를 util로 분리
